refactor(be): CQRS 스타일 도입하여 조회 작업 클래스 추가
