+++ backend/src/main/java/com/deliveranything/domain/order/service/OrderService.java
@@ -20,7 +20,6 @@ import com.deliveranything.domain.order.event.sse.seller.OrderCanceledForSellerE
 import com.deliveranything.domain.order.event.sse.seller.OrderPaidForSellerEvent;
 import com.deliveranything.domain.order.event.sse.seller.OrderPreparingForSellerEvent;
 import com.deliveranything.domain.order.event.sse.seller.OrderStatusChangedForSellerEvent;
-import com.deliveranything.domain.order.service.fetcher.OrderFetcher;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.ApplicationEventPublisher;
@@ -33,17 +32,17 @@ import org.springframework.transaction.annotation.Transactional;
 public class OrderService {
 
   private final ApplicationEventPublisher eventPublisher;
-  private final OrderFetcher orderFetcher;
+  private final OrderQueryService orderQueryService;
 
   @Transactional
   public void processPaymentCompletion(String merchantUid) {
-    Order order = orderFetcher.findByMerchantIdOrThrow(merchantUid);
+    Order order = orderQueryService.findByMerchantIdOrThrow(merchantUid);
     eventPublisher.publishEvent(OrderPaymentSucceededEvent.fromOrder(order));
   }
 
   @Transactional
   public void processStockCommitted(Long orderId) {
-    Order order = orderFetcher.findByIdOrThrow(orderId);
+    Order order = orderQueryService.findByIdOrThrow(orderId);
     order.updateStatus(OrderStatus.PENDING);
 
     eventPublisher.publishEvent(OrderPaidForCustomerEvent.fromOrder(order));
@@ -52,7 +51,7 @@ public class OrderService {
 
   @Transactional
   public void processOrderTransmitted(Long orderId) {
-    Order order = orderFetcher.findByIdOrThrow(orderId);
+    Order order = orderQueryService.findByIdOrThrow(orderId);
     order.updateStatus(OrderStatus.PREPARING);
 
     eventPublisher.publishEvent(OrderPreparingForCustomerEvent.fromOrder(order));
@@ -61,13 +60,13 @@ public class OrderService {
 
   @Transactional
   public void processPaymentFailure(String merchantUid) {
-    Order order = orderFetcher.findByMerchantIdOrThrow(merchantUid);
+    Order order = orderQueryService.findByMerchantIdOrThrow(merchantUid);
     eventPublisher.publishEvent(OrderPaymentFailedEvent.fromOrder(order));
   }
 
   @Transactional
   public void processStockReleased(Long orderId) {
-    Order order = orderFetcher.findByIdOrThrow(orderId);
+    Order order = orderQueryService.findByIdOrThrow(orderId);
     order.updateStatus(OrderStatus.PAYMENT_FAILED);
 
     eventPublisher.publishEvent(OrderPaymentFailedForCustomerEvent.fromOrder(order));
@@ -75,7 +74,7 @@ public class OrderService {
 
   @Transactional
   public void processPaymentCancelSuccess(String merchantUid, Publisher publisher) {
-    Order order = orderFetcher.findByMerchantIdOrThrow(merchantUid);
+    Order order = orderQueryService.findByMerchantIdOrThrow(merchantUid);
 
     if (publisher == Publisher.CUSTOMER) {
       order.updateStatus(OrderStatus.CANCELED);
@@ -88,7 +87,7 @@ public class OrderService {
 
   @Transactional
   public void processStockReplenished(Long orderId) {
-    Order order = orderFetcher.findByIdOrThrow(orderId);
+    Order order = orderQueryService.findByIdOrThrow(orderId);
 
     eventPublisher.publishEvent(OrderCanceledForCustomerEvent.fromOrder(order));
     eventPublisher.publishEvent(OrderCanceledForSellerEvent.fromOrder(order));
@@ -96,7 +95,7 @@ public class OrderService {
 
   @Transactional
   public void processPaymentCancelFailed(String merchantUid) {
-    Order order = orderFetcher.findByMerchantIdOrThrow(merchantUid);
+    Order order = orderQueryService.findByMerchantIdOrThrow(merchantUid);
     order.updateStatus(OrderStatus.CANCEL_FAILED);
 
     eventPublisher.publishEvent(OrderCancelFailedForCustomerEvent.fromOrder(order));
@@ -105,7 +104,7 @@ public class OrderService {
 
   @Transactional
   public void processDeliveryPickedUp(Long orderId) {
-    Order order = orderFetcher.findByIdOrThrow(orderId);
+    Order order = orderQueryService.findByIdOrThrow(orderId);
     order.updateStatus(OrderStatus.DELIVERING);
 
     eventPublisher.publishEvent(OrderStatusChangedForCustomerEvent.fromOrder(order));
@@ -114,7 +113,7 @@ public class OrderService {
 
   @Transactional
   public void processDeliveryCompleted(Long orderId, Long riderId, Long sellerId) {
-    Order order = orderFetcher.findByIdOrThrow(orderId);
+    Order order = orderQueryService.findByIdOrThrow(orderId);
     order.updateStatus(OrderStatus.COMPLETED);
 
     eventPublisher.publishEvent(OrderCompletedEvent.fromOrder(order, riderId, sellerId));
@@ -124,7 +123,7 @@ public class OrderService {
 
   @Transactional(readOnly = true)
   public void processStockReserved(Long orderId) {
-    Order order = orderFetcher.findByIdOrThrow(orderId);
+    Order order = orderQueryService.findByIdOrThrow(orderId);
     log.info("주문 재고 홀드 됨. 클라이언트에게 주문 생성 관련 정보 전달.");
 
     eventPublisher.publishEvent(OrderCreatedForCustomerEvent.fromOrder(order));
@@ -134,7 +133,7 @@ public class OrderService {
   public void processStockReserveFailed(Long orderId, String reason) {
     log.info("주문 [{}] 취소 처리 시작. 사유: 재고 예약 실패 ({})", orderId, reason);
 
-    Order order = orderFetcher.findByIdOrThrow(orderId);
+    Order order = orderQueryService.findByIdOrThrow(orderId);
     order.cancel(reason);
     eventPublisher.publishEvent(OrderCreateFailedForCustomerEvent.fromOrder(order));
