+++ backend/src/main/java/com/deliveranything/domain/order/service/PaymentOrderService.java
@@ -4,7 +4,6 @@ import com.deliveranything.domain.order.entity.Order;
 import com.deliveranything.domain.order.enums.Publisher;
 import com.deliveranything.domain.order.event.OrderCancelEvent;
 import com.deliveranything.domain.order.event.OrderPaymentRequestedEvent;
-import com.deliveranything.domain.order.service.fetcher.OrderFetcher;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.ApplicationEventPublisher;
@@ -17,13 +16,13 @@ import org.springframework.transaction.annotation.Transactional;
 public class PaymentOrderService {
 
   private final ApplicationEventPublisher eventPublisher;
-  private final OrderFetcher orderFetcher;
+  private final OrderQueryService orderQueryService;
 
   @Transactional
   public void payOrder(String merchantUid, String paymentKey) {
     log.info("고객의 결제 시도 - 주문 번호: {} & PG 사 결제 키: {}", merchantUid, paymentKey);
 
-    Order order = orderFetcher.findByMerchantIdOrThrow(merchantUid);
+    Order order = orderQueryService.findByMerchantIdOrThrow(merchantUid);
     order.isPayable();
 
     eventPublisher.publishEvent(
@@ -32,7 +31,7 @@ public class PaymentOrderService {
 
   @Transactional
   public void cancelOrder(Long orderId, String cancelReason) {
-    Order order = orderFetcher.findByIdOrThrow(orderId);
+    Order order = orderQueryService.findByIdOrThrow(orderId);
     order.cancellationRequest(cancelReason);
 
     eventPublisher.publishEvent(OrderCancelEvent.from(order, cancelReason, Publisher.CUSTOMER));
